Update edition and dependencies
- Cargo clippy fixes

+++ src/poly/resultant.rs
@@ -72,9 +72,9 @@ impl<F: EuclideanDomain> UnivariatePolynomial<F> {
         let mut rho = self.ring.one();
         let mut den = self.ring.one();
         for k in 1..lcs.len() {
-            let mut deg = lcs[k as usize - 1].1 as i64 - lcs[k as usize].1 as i64;
+            let mut deg = lcs[k - 1].1 as i64 - lcs[k].1 as i64;
             for l in k..lcs.len() - 1 {
-                deg *= 1 - (lcs[l as usize].1 as i64 - lcs[l as usize + 1].1 as i64) as i64;
+                deg *= 1 - (lcs[l].1 as i64 - lcs[l + 1].1 as i64) as i64;
             }
 
             if deg > 0 {
